+++ shared.c
@@ -2661,7 +2661,7 @@ va_list         vl;
     gettimeofday(&tv, NULL);
     sprintf ((char *)s,
             "%6.6ld" "." "%6.6ld %4.4X:",
-            tv.tv_sec, tv.tv_usec, dev ? dev->devnum : 0);
+            tv.tv_sec, (long)(tv.tv_usec), dev ? dev->devnum : 0);
     vsnprintf ((char *)s + strlen(s), sizeof(s) - strlen(s),
             msg, vl);
     if (dt)
